feat: use worker_threads to watch files
